--- rodan/jobs/resource_distributor/__init__.py
@@ -1,3 +1,0 @@
-__version__ = "1.2.0"
-
-from .base import ResourceDistributor
